+++ src/main/java/com/freddy/verticaltabs/TabDefinition.java
@@ -2,20 +2,13 @@ package com.freddy.verticaltabs;
 
 final class TabDefinition
 {
-    private final int index;
     private final String name;
     private final String fallback;
 
-    TabDefinition(int index, String name, String fallback)
+    TabDefinition(String name, String fallback)
     {
-        this.index = index;
         this.name = name;
         this.fallback = fallback;
-    }
-
-    int getIndex()
-    {
-        return index;
     }
 
     String getName()